requirements.txt
pyhtml2pdf==0.0.7
boto3==1.28.80
PyPDF2==2.12.1
python-dotenv==1.0.0